+++ web/src/Pages/Home.js
@@ -1,24 +1,15 @@
 import logo from '../logo.png';
 import '../App.css';
+import './Home.css';
 import HomeBanner from '../Components/HomeBanner';
 import React from 'react';
 
 class Home extends React.Component {
   render() {
     return (
-    <div className="App-header">
+    <div className="pageWrapper">
       <img src={logo} className="App-logo" alt="logo" />
-      <p>
-        goscrobble.com
-      </p>
-      <a
-        className="App-link"
-        href="https://gitlab.com/idanoo/go-scrobble"
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        gitlab.com/idanoo/go-scrobble
-      </a>
+      <p className="homeText">Go-Scrobble is an open source music scrobbling service written in Go and React.</p>
       <HomeBanner />
     </div>
     );